refactor(ui): split graph axis helpers into result_graph_axes.go
Keeps result_graph.go under the 200 LOC file cap.
