+++ internal/ui/result_graph_axes.go
@@ -1,0 +1,102 @@
+package ui
+
+import (
+	"fmt"
+	"strings"
+)
+
+// axisPipe returns the left (┤/│) or right (├/│) axis join for a row; tick rows
+// carry the horizontal stub, plain rows a continuous vertical line.
+func axisPipe(left, tick bool) string {
+	if tick {
+		if left {
+			return "┤"
+		}
+		return "├"
+	}
+	return "│"
+}
+
+// wpmAxisLabel renders the left Y-axis WPM tick (top=max, mid=max/2, bottom=0).
+func wpmAxisLabel(cr, chartH, midRow int, maxWPM float64) string {
+	switch {
+	case cr == 0:
+		return fmt.Sprintf("%4.0f", maxWPM)
+	case cr == chartH-1:
+		return "   0"
+	case cr == midRow:
+		return fmt.Sprintf("%4.0f", maxWPM/2)
+	default:
+		return strings.Repeat(" ", 4)
+	}
+}
+
+// errAxisLabel renders the right Y-axis error tick (top=max, mid=max/2, bottom=0).
+func errAxisLabel(cr, chartH, midRow, maxErr int) string {
+	switch {
+	case cr == 0:
+		return fmt.Sprintf("%-3d", maxErr)
+	case cr == chartH-1:
+		return "0  "
+	case cr == midRow:
+		return fmt.Sprintf("%-3d", maxErr/2)
+	default:
+		return strings.Repeat(" ", 3)
+	}
+}
+
+// clampInt pins v into [lo,hi].
+func clampInt(v, lo, hi int) int {
+	if v < lo {
+		return lo
+	}
+	if v > hi {
+		return hi
+	}
+	return v
+}
+
+// minMax returns the minimum and maximum values in vals. Assumes len(vals) > 0.
+func minMax(vals []float64) (min, max float64) {
+	min, max = vals[0], vals[0]
+	for _, v := range vals[1:] {
+		if v < min {
+			min = v
+		}
+		if v > max {
+			max = v
+		}
+	}
+	return min, max
+}
+
+// xAxisLabels builds a compact x-axis label row, placing second markers at
+// intervals of 4 (or every second for short sequences) to keep it readable.
+// Returns a plain width-n string; the caller applies styling.
+func xAxisLabels(n int) string {
+	if n == 0 {
+		return ""
+	}
+	step := 4
+	if n <= 4 {
+		step = 1
+	}
+	var parts []string
+	lastEnd := 0
+	for i := 0; i < n; i += step {
+		label := fmt.Sprintf("%d", i)
+		if i+len(label) <= n {
+			for lastEnd < i {
+				parts = append(parts, " ")
+				lastEnd++
+			}
+			parts = append(parts, label)
+			lastEnd += len(label)
+		}
+	}
+	for lastEnd < n {
+		parts = append(parts, " ")
+		lastEnd++
+	}
+	return strings.Join(parts, "")
+}
